requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv
pyautogui
pyperclip
pywin32
requests
numpy
scipy
Pillow
